Move macOS MacCentralProcessor, MacFirmware, and MacHWDiskStore superclasses to oshi-common; add JNA suffix to remaining mac hardware classes (#3148)
